Miscellaneous fixes
+ Added documentation for each command
+ Added PostRun to close DB for each action that opens it
+ Added args check for commands that need arguments

+++ cmd/delete.go
@@ -2,6 +2,7 @@ package cmd
 
 import (
 	"fmt"
+	"os"
 
 	"github.com/MattRighetti/passwdvault/configuration"
 	db "github.com/MattRighetti/passwdvault/database"
@@ -11,11 +12,16 @@ import (
 var deleteCmd = &cobra.Command{
 	Use:   "delete [PASSWORD_ID]",
 	Short: "Deletes password",
-	Long:  "examples here...",
+	Long:  "Delete will search for the password of the passed pasword identifier and, if present, it will delete it",
 	PreRunE: func(cmd *cobra.Command, args []string) error {
 		return configuration.InitCriticalData()
 	},
 	Run: func(cmd *cobra.Command, args []string) {
+		if len(args) == 0 {
+			cmd.Help()
+			os.Exit(0)
+		}
+
 		if db.IsPresent(args[0]) {
 			err := db.Delete(args[0])
 			if err != nil {
@@ -26,6 +32,9 @@ var deleteCmd = &cobra.Command{
 		} else {
 			fmt.Printf("No password stored with key %s\n", args[0])
 		}
+	},
+	PostRun: func(cmd *cobra.Command, args []string) {
+		configuration.CloseDb()
 	},
 }
